+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 # Changelog
 
-## Unreleased
+<!-- next-header -->
+
+## Unreleased - FutureDate
 ### Changed
  - The `pes::ElementaryStreamConsumer` type and its methods are now parameterised to gain access to the context object
-   you provided to the demultiplexer.
+   you provided to the demultiplexer.  Thanks @fkaa.
  - `RegistrationDescriptor::format_identifier()` return type changed from u32 to a value of the `FormatIdentifier` enum
    from the [smptera-format-identifiers-rust](https://crates.io/crates/smptera-format-identifiers-rust) crate.  Also
    `RegistrationDescriptor` fields are no longer public.