+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 <!-- next-header -->
 
 ## Unreleased - FutureDate
+
+## 0.14.0 - 2021-04-11
 ### Changed
  - The `pes::ElementaryStreamConsumer` type and its methods are now parameterised to gain access to the context object
    you provided to the demultiplexer.  Thanks @fkaa.